refactor: TicketUI 코드를 MVP로 변환
