+++ app/src/main/java/woowacourse/movie/view/ticket/TicketView.kt
@@ -1,0 +1,10 @@
+package woowacourse.movie.view.ticket
+
+import woowacourse.movie.model.ticket.Ticket
+import woowacourse.movie.view.model.TicketData
+
+interface TicketView {
+    fun getTicketData(): TicketData
+
+    fun initTicketUI(ticket: Ticket)
+}
